# specs/001-clinic-scheduling/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Patient Scheduling API
  version: 0.1.0
  description: >
    Tenant-scoped REST API for clinic patient scheduling. Authentication is via a
    server-side session cookie that carries the user's center scope. Every non-super-admin
    endpoint is scoped to the caller's center and fails closed (403, no data) if scope
    cannot be established (Constitution Principle II). All times are UTC (RFC 3339).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    date:
      name: date
      in: query
      schema: { type: string, format: date }
    doctorId:
      name: doctor_id
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: double_booking }
        message: { type: string }
...
    Center:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string, example: Asia/Riyadh }
        grid_minutes: { type: integer, default: 15 }
        status: { type: string, enum: [active, suspended] }
    CenterCreate:
      type: object
      required: [name, admin_email, admin_temp_password]
      properties:
        name: { type: string }
        timezone: { type: string, default: Asia/Riyadh, example: Asia/Riyadh }
        grid_minutes: { type: integer, default: 15 }
        admin_email: { type: string, format: email }
        admin_temp_password: { type: string }
        admin_role: { type: string, enum: [center_admin, doctor, reception], default: center_admin }
    CenterUpdate:
      type: object
      required: [timezone]
      properties:
        timezone: { type: string, example: Asia/Riyadh }
...
    AvailabilityTemplate:
      type: object
      required: [doctor_id, weekday, start_local, end_local]
      properties:
        id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        clinic_id: { type: string, format: uuid, nullable: true }
        weekday: { type: integer, minimum: 0, maximum: 6 }
        start_local: { type: string, example: "09:00" }
        end_local: { type: string, example: "13:00" }
    AvailabilityException:
      type: object
      required: [doctor_id, date, kind]
      properties:
        id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        clinic_id: { type: string, format: uuid, nullable: true }
        date: { type: string, format: date }
        kind: { type: string, enum: [block, override, extra] }
        start_local: { type: string, nullable: true }
        end_local: { type: string, nullable: true }
        reason: { type: string, nullable: true }
    BookableInterval:
      type: object
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
    Clinic:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        is_active: { type: boolean }
    ClinicCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
    ClinicUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 80, nullable: true }
        is_active: { type: boolean, nullable: true }
...
  responses:
    Forbidden:
      description: Not permitted / out of center scope (default-deny)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Integrity conflict (e.g., double_booking, slot_taken, invalid_transition)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Validation error (e.g., outside_availability, off_grid, patient_conflict)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - sessionCookie: []
paths:
...
  /centers:
    get:
      tags: [centers]
      summary: List all centers (super-admin)
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Center' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [centers]
      summary: Create a center and its first admin (super-admin)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CenterCreate' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Center' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /centers/{centerId}:
    put:
      tags: [centers]
      summary: Update a center's time zone (super-admin)
      parameters: [ { name: centerId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CenterUpdate' } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Center' } } } }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { description: Not found }, '422': { $ref: '#/components/responses/Unprocessable' } }
  /centers/{centerId}/status:
    put:
      tags: [centers]
      summary: Suspend or reactivate a center (super-admin)
      parameters: [ { name: centerId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { status: { type: string, enum: [active, suspended] } } } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Center' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }

  # ---- clinics (working units; managed by center-admin) ----
  /clinics:
    get:
      tags: [clinics]
      summary: List clinics (working units) in caller's center
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Clinic' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [clinics]
      summary: Add a clinic (center-admin)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClinicCreate' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Clinic' } } } }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
  /clinics/{clinicId}:
    put:
      tags: [clinics]
      summary: Rename or activate/deactivate a clinic (center-admin)
      parameters: [ { name: clinicId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClinicUpdate' } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Clinic' } } } }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { description: Not found }, '409': { $ref: '#/components/responses/Conflict' } }
    delete:
      tags: [clinics]
      summary: Remove a clinic (center-admin); 409 if it has appointments
      parameters: [ { name: clinicId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Deleted }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { description: Not found }, '409': { $ref: '#/components/responses/Conflict' } }

  # ---- users (center-admin) ----
...
  /availability/templates:
    get:
      tags: [availability]
      summary: List a doctor's weekly templates
      parameters: [ { $ref: '#/components/parameters/doctorId' } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AvailabilityTemplate' } } } } } }
    post:
      tags: [availability]
      summary: Add a weekly template entry (doctor or center-admin)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityTemplate' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityTemplate' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /availability/templates/{templateId}:
    delete:
      tags: [availability]
      summary: Remove a weekly template entry (doctor or center-admin)
      parameters: [ { name: templateId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Deleted }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { description: Not found } }
  /availability/exceptions:
    get:
      tags: [availability]
      summary: List a doctor's date-specific exceptions (optionally for one date)
      parameters: [ { $ref: '#/components/parameters/doctorId' }, { name: date, in: query, required: false, schema: { type: string, format: date } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AvailabilityException' } } } } } }
    post:
      tags: [availability]
      summary: Add a date-specific exception (block / override / extra)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityException' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AvailabilityException' } } } } }
  /availability/exceptions/{exceptionId}:
    delete:
      tags: [availability]
      summary: Remove a date-specific exception (doctor or center-admin)
      parameters: [ { name: exceptionId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Deleted }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { description: Not found } }
  /availability/slots:
    get:
      tags: [availability]
      summary: Resolved bookable intervals for a doctor on a date (template + exceptions)
      parameters: [ { $ref: '#/components/parameters/doctorId' }, { $ref: '#/components/parameters/date' } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/BookableInterval' } } } } } }

  # ---- requests (doctor -> reception) ----
...